// libs/core/src/lib/enums/role.ts
export enum Role {
  VALIDATOR_1 = 'Validator1', //First level validator in field office
  VALIDATOR_2 = 'Validator2', //Second level validator in field office
  VALIDATOR__1 = 'Validator 1',
  VALIDATOR__2 = 'Validator 2',
  ADMIN = 'admin',
  ADMIN_TNX = 'Admin',
  CSR = 'CSR',
  EST_ADMIN = 'estadmin',
  VALIDATOR = 'Validator', //Only Validator in Gosi Online
  USER = 'user',
  FC_APPROVER = 'FinanceControl',
  FC_APPROVER_ANNUITY = 'FinanceController',
  OH_FCAPPROVER = 'Financial Controller - OH',
  OH_FCAPPROVER1 = 'FinancialController-OH',
  OH_FC_APPROVER = 'FinancialController-OH',
  CNT_FC_APPROVER = 'FCApprover',
  EST_ADMIN_OH = 'Admin',
  WORK_INJURIES_AND_OCCUPATIONAL_DISEASES_DOCTOR = 'Work Injuries and Occupational Diseases Doctor',
  CONTRIBUTOR = 'contributor',
  CONTRIBUTOR_SANED = 'Contributor',
  GDS = 'GeneralDirectorofOperationsandInsuranceServices',
  OH_GDS = 'GeneralDirectorofInsuranceServicesandOperations',
  DOCTOR = 'Doctor',
  AUDITOR = 'Auditor',
  SM = 'SectionManager', // new name for section manager is BOSH
  BOSH = 'BenefitOperationSectionHead', // new name for section manager
  VCSANED = 'Validationcommittee(SANEDAppeal)',
  VC = 'ValidationCommittee',
  VCAPPEAL = 'Validation committee (SANED Appeal)',
  OWNER = 'Owner',
  FO_DIRECTOR = 'Headofclaimmanagementdepartment',
  GDISO = 'GDISO',
  GDES = 'GDES',
  CSB_DIRECTOR = 'Customerserviceandbranchesgeneraldirector',
  INSURANCE_OP_HEAD = 'InsuranceOperationsManager',
  INSURANCE_DIRECTOR = 'GeneralDirectorofInsuranceServicesandOperations',
  OH_INSURANCE_DIRECTOR = 'General Director of Insurance Services and Operations',
  OH_OFFICER = 'OccupationalHazardOperationsOfficer',
  OH_DOCTOR = 'WorkInjuriesandOccupationalDiseasesDoctor',
  MC_OFFICER = 'MedicalCodingOfficer',
  APPEAL_MEDICAL = 'AppealMedicalServicesOfficer',
  CLM_MGR = 'ClaimsDepartmentManager',
  OCCUPATIONAL_HAZARD_OPERATIONS_OFFICER = 'Occupational Hazard Operations Officer',
  BOARD_OFFICER = 'MedicalCostController',
  BRANCH_ADMIN = 'BranchAdmin',
  OH_ADMIN = 'OccupationalHazardAdmin',
  OH_AUDITOR = 'Medical Auditor',
  FC_CONTROLLER = 'FinancialController',
  SAFETY_HEALTH_ENGINEER = 'Occupational Health & Safety Inspector',
  SUPER_ADMIN = 'BranchesAccountManager',
  MS_OFFICER = 'MedicalServicesOfficer',
  CUSTOMER_CARE_OFFICER = 'CustomerCareOfficer',
  CUSTOMER_CARE_SENIOR_OFFICER = 'CustomerCareSeniorOfficer',
  DEPARTMENT_HEAD = 'ComplaintManager',
  CLERK = 'Clerk',
  DEPARTMENT_CLERK = 'ComplaintClerk',
  PERSON = 'Person',
  COMMITEE_MEMBER = 'ViolationCommitteeMember',
  COMMITEE_HEAD = 'ViolationCommitteeHead',
  CUSTOMER_SERVICE_SUPERVISOR = 'Customer Service Supervisor',
  CUSTOMER_SERVICE_REPRESENTATIVE = 'Customer Service Representative',
  EXCEPTIONAL_WAIVER_OFFICER = 'ExceptionalWaiverOfficer',
  MEDICAL_MANAGER = 'MedicalServicesDepartmentManager',
  MEDICAL_OFFICER = 'MedicalServiceOfficer',
  BRANCH_MANAGER = 'BranchManager',
  COMPLIANCE_MANAGER = 'GeneralManagerofComplianceandInsurancePrevention',
  FINANCIAL_CONTROLLER_ANN_ONE = 'Financial Controller - ANN 1',
  FINANCIAL_CONTROLLER_ANN_TWO = 'Financial Controller - ANN 2',
  INSURANCE_BENEFIT_OPERATIONAL_OFFICER_TWO = 'Insurance Benefit Operational Officer 2',
  BENEFIT_SEARCH_READ = 'Benefits Search - Read',
  INSURANCE_BENEFIT_OPERATIONAL_OFFICER_ONE = 'Insurance Benefit operational officer 1',
  BACKDATE_ENGAGEMENT_VALIDATOR = 'Backdated Engagement Validator',
  AOV_REVIEWER = 'ISreviewerviolation',
  AOV_APPROVER_1 = 'AOVApprover1',
  AOV_APPROVER_2 = 'AOVApprover2',
  AOV_APPROVER_3 = 'AOVApprover3',
  IS_REVIEWER_VIOLATION = 'ISreviewerviolation',
  Legal_Auditor_Violation = 'legalauditorviolation',
  Legal_Reviewer_Violation = 'legalreviewerviolation',
  Legal_Manger_Violation = 'legalmangerviolation',
  Executor_Violation = 'executorviolation',
  Individuals_Preparation_Specialist_Violation = 'Individualspreparationspecialistviolation',
  SANED_VALIDATION_COMMITTEE = 'Saned Validation Committee',
  AOV_EXECUTER = 'AOVExecuter',
  IS_reviewer_private = 'IS reviewer private',
  ESTABLISHMENT_ADMIN = 'EstablishmentAdmin',
  MB_OFFICER = 'Medical Board Officer',
  MBO = 'MedicalBoardOfficer',
  GOSI_DOCTOR = 'GOSIDOCTOR',
  GOSI_DOCTOR1 = 'GOSI DOCTOR',
  HO_DOCTOR = 'HODOCTOR',
  HO_OFFICER = 'HO DOCTOR',
  CONTRACT_DOCTOR = 'ContractDoctor',
  HEAD_GOSI_DOCTOR = 'Head Office Gosi Doctor',
  HEAD_OFFICE_GOSI_DOCTOR = 'HeadOfficeGosiDoctor',
  GOSI_DOCTOR_CAPS = 'Work Injuries and Occupational Diseases Doctor',
  MEDICAL_BOARD_OFFICER = 'Medical Services Officer',
  APPEAL_MEDICAL_BOARD_OFFICER = 'Appeal Medical Board Officer',
  APPEAL_OFFICER = 'AppealMedicalBoardOfficer',
  CONTRACTED_DOCTOR = 'Contracted Doctor',
  CONTRACT_DOC = 'ContractedDoctor',
  MEDICAL_MANAGER_SPACE = 'Medical Services Department Manager',
  ESTABLISHMENT = 'Establishment',
  ADMINPOOL = 'ADMINPOOL'
}